reorganize containers
