+++ components/containers/MainContainer.jsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export default MainContainer = styled.main`
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  width: 100%;
+`;
