+++ components/containers/TitleContainer.jsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export default TitleContainer = styled.div`
+  display: flex;
+  margin: 0 0 1rem 0;
+  align-items: center;
+  justify-content: space-evenly;
+
+  h1 {
+    font-size: 3rem;
+    margin: 0;
+  }
+`;
