+++ components/plots/LinePlot.jsx
@@ -2,7 +2,7 @@ import uPlot from "uplot";
 import React from "react";
 import styled from "styled-components";
 
-import { ResizableContainer } from "../containers";
+import ResizableContainer from "../containers/ResizableContainer";
 import { Colors } from "../../lib/constants";
 import { tooltipsPlugin } from "../../lib/chartPlugins";
