+++ pages/[page].jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useRouter } from "next/router";
 
-import { MainContainer } from "../components/containers";
+import MainContainer from "../components/containers/MainContainer";
 import {
   getAllPosts,
   markdownToHtml,
