+++ pages/tag/[tag].jsx
@@ -2,7 +2,7 @@ import React from "react";
 import styled from "styled-components";
 import { useRouter } from "next/router";
 
-import { MainContainer } from "../../components/containers";
+import MainContainer from "../../components/containers/MainContainer";
 import { getAllPosts, markdownToHtml, POST_FIELDS } from "../../lib/blogs";
 import { sizeImage } from "../../lib/util";
